backend-django/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
virtualenv==20.19.0